script/requirements.txt
antlr4-python3-runtime==4.7.2
certifi==2019.11.28
chardet==3.0.4
cybox==2.1.0.20
elasticsearch==7.1.0
idna==2.8
lxml==4.4.2
mixbox==1.0.3
openioc-to-stix==1.4.0
ordered-set==3.1.1
ordereddict==1.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
requests==2.22.0
simplejson==3.17.0
six==1.13.0
stix==1.2.0.7
stix-validator==2.5.0
stix2==1.1.0
stix2-patterns==1.2.1
stix2-slider==2.0.0
stixmarx==1.0.6
tornado==6.0.3
urllib3==1.25.7
xlrd==1.2.0
